[REFACTOR] 공유 프로세스 (#594)
