+++ app/src/main/java/org/sopt/havit/ui/share/add_category/EnterCategoryTitleFragment.kt
@@ -1,4 +1,4 @@
-package org.sopt.havit.ui.share
+package org.sopt.havit.ui.share.add_category
 
 import android.os.Bundle
 import android.view.View
