+++ app/src/main/java/org/sopt/havit/ui/share/edit_title/EditTitleFragment.kt
@@ -1,4 +1,4 @@
-package org.sopt.havit.ui.share
+package org.sopt.havit.ui.share.edit_title
 
 import android.os.Bundle
 import android.view.LayoutInflater
