+++ app/src/main/res/layout/fragment_contents_summery.xml
@@ -18,7 +18,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.share.ContentsSummeryFragment">
+        tools:context=".ui.share.contents_summery.ContentsSummeryFragment">
 
         <!-- Toolbar-->
         <androidx.constraintlayout.widget.ConstraintLayout
