+++ app/src/main/res/layout/fragment_edit_title.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.share.EditTitleFragment">
+    tools:context=".ui.share.edit_title.EditTitleFragment">
 
     <!-- toolbar -->
     <androidx.constraintlayout.widget.ConstraintLayout
