+++ app/src/main/res/layout/fragment_enter_category_title.xml
@@ -21,7 +21,7 @@
         android:id="@+id/enterCategoryTitleFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.share.EnterCategoryTitleFragment">
+        tools:context=".ui.share.add_category.EnterCategoryTitleFragment">
 
         <!-- Toolbar -->
         <androidx.constraintlayout.widget.ConstraintLayout
