+++ app/src/main/res/layout/fragment_picker.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/ellipse_white_radius_20"
-    tools:context=".ui.share.PickerFragment">
+    tools:context=".ui.share.notification.PickerFragment">
 
     <ImageView
         android:id="@+id/iv_alarm"
