+++ app/src/main/res/layout/fragment_set_notification.xml
@@ -13,7 +13,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.share.SetNotificationFragment">
+        tools:context=".ui.share.notification.SetNotificationFragment">
 
         <!-- 툴바 -->
         <androidx.constraintlayout.widget.ConstraintLayout
